requirements.txt
flask
Flask-SQLAlchemy==1.3.23
flask-login